+++ 6DOF_object_detector/include/utils.h
@@ -19,15 +19,15 @@
 
 struct Line {
 
-    Eigen::Vector3f direction;
-    Eigen::Vector3f point;
+    Eigen::Vector3d direction;
+    Eigen::Vector3d point;
 };
 
 struct Plane {
-    Eigen::Vector4f coefficients;
-    Eigen::Vector3f point;
-    Eigen::Vector3f getNormal(){
-        Eigen::Vector3f normal = Eigen::Vector3f(coefficients[0], coefficients[1], coefficients[2]);
+    Eigen::Vector4d coefficients;
+    Eigen::Vector3d point;
+    Eigen::Vector3d getNormal(){
+        Eigen::Vector3d normal = Eigen::Vector3d(coefficients[0], coefficients[1], coefficients[2]);
         return normal;
     }
 
@@ -69,32 +69,32 @@ bool isInsideKernel2D(float x, float y, float cx, float cy , float radius);
 // Calculate PCA over mask
 void calcPCA(cv::Mat &img_mask, cv::Point2f &meanPoint, cv::Size2f &dimension, float &rotAngle);
 
-Eigen::Matrix3f quaternionToMatrix(Eigen::Quaternionf &q);
+Eigen::Matrix3d quaternionToMatrix(Eigen::Quaterniond &q);
 
-// Eigen::Quaternionf logQuaternion(Eigen::Quaternionf q);
+// Eigen::Quaterniond logQuaternion(Eigen::Quaterniond q);
 
-//Eigen::Quaternionf expQuaternion(Eigen::Quaternionf q);
+//Eigen::Quaterniond expQuaternion(Eigen::Quaterniond q);
 
 // generalized quaternion interpolation
-Eigen::Quaternionf quatInterp(const std::vector<Eigen::Quaternionf>& rotation);
+Eigen::Quaterniond quatInterp(const std::vector<Eigen::Quaterniond>& rotation);
 
-void selectConvexHull( const cv::Mat& img_rgb, const pcl::PointCloud< pcl::PointXYZRGB >::Ptr& cloud, Eigen::Matrix4d& referenceTransform, std::vector< Eigen::Vector3f, Eigen::aligned_allocator< Eigen::Vector3f > >& convexHull_ ) ;
+void selectConvexHull( const cv::Mat& img_rgb, const pcl::PointCloud< pcl::PointXYZRGB >::Ptr& cloud, Eigen::Matrix4d& referenceTransform, std::vector< Eigen::Vector3d, Eigen::aligned_allocator< Eigen::Vector3d > >& convexHull_ ) ;
 
-void selectPlane( const cv::Mat& img_rgb, const pcl::PointCloud< pcl::PointXYZRGB >::Ptr& cloud, Eigen::Matrix4d& referenceTransform, std::vector< Eigen::Vector3f, Eigen::aligned_allocator< Eigen::Vector3f > > &convexHull_, Plane &table_plane ) ;
+void selectPlane( const cv::Mat& img_rgb, const pcl::PointCloud< pcl::PointXYZRGB >::Ptr& cloud, Eigen::Matrix4d& referenceTransform, std::vector< Eigen::Vector3d, Eigen::aligned_allocator< Eigen::Vector3d > > &convexHull_, Plane &table_plane ) ;
 
 void getObjectPointCloud( const pcl::PointCloud< pcl::PointXYZRGB >::ConstPtr& cloud, float minHeight, float maxHeight,
-                          std::vector< Eigen::Vector3f, Eigen::aligned_allocator< Eigen::Vector3f > > convexHull, Plane &table_plane, Eigen::Vector3f turnTable_center, pcl::PointCloud<pcl::PointXYZRGB>::Ptr &objectCloud  ) ;
+                          std::vector< Eigen::Vector3d, Eigen::aligned_allocator< Eigen::Vector3d > > convexHull, Plane &table_plane, Eigen::Vector3d turnTable_center, pcl::PointCloud<pcl::PointXYZRGB>::Ptr &objectCloud  ) ;
 
-Eigen::Vector3f getTurnTableCenter( const cv::Mat& img_rgb, const pcl::PointCloud< pcl::PointXYZRGB >::Ptr& cloud, Eigen::Matrix4d& referenceTransform, Plane &table_plane ) ;
+Eigen::Vector3d getTurnTableCenter( const cv::Mat& img_rgb, const pcl::PointCloud< pcl::PointXYZRGB >::Ptr& cloud, Eigen::Matrix4d& referenceTransform, Plane &table_plane ) ;
 
 void printScore(cv::Mat &img, string &objectName, float score, cv::Point2f &pt, bool print_score );
 
-void get3DBoundingBox(pcl::PointCloud<pcl::PointXYZRGB>::Ptr &cloud, std::vector< Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f> >& transformationMatrixOC, std::vector< cv::Point3f> &cg, cv::Point3f &bbSize);
+void get3DBoundingBox(pcl::PointCloud<pcl::PointXYZRGB>::Ptr &cloud, std::vector< Eigen::Matrix4d, Eigen::aligned_allocator<Eigen::Matrix4d> >& transformationMatrixOC, std::vector< cv::Point3f> &cg, cv::Point3f &bbSize);
 
-void create3DBB(cv::Point3f &bbSize, Eigen::Matrix4f &transformationMatrixOC , cv::Size2f &img_size, std::vector< cv::Point2f > &imagePoints);
+void create3DBB(cv::Point3f &bbSize, Eigen::Matrix4d &transformationMatrixOC , cv::Size2f &img_size, std::vector< cv::Point2f > &imagePoints);
 
 void createWireFrame (cv::Mat &img, std::vector<cv::Point2f> &vertices);
 
 void  getLine( pcl::PointXYZRGB  &p,  Line&line);
 
-void getLinePlaneIntersection(Line &line, Plane &plane, Eigen::Vector3f &ptIntersection);
+void getLinePlaneIntersection(Line &line, Plane &plane, Eigen::Vector3d &ptIntersection);
